requirements.txt
discord
disnake-pagination
discordpy-antispam
python-dotenv